# internal/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cannon Content API",
    "version": "1.0.0",
    "description": "Headless read-only content and account self-service API for Cannon CMS. All endpoints require the X-Cannon-API-Key header. User-specific visibility and account routes also accept Authorization: Bearer <JWT>."
  },
  "servers": [{ "url": "/api/v1" }],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Cannon-API-Key",
        "description": "Application credential issued in Admin → API → Credentials"
      },
      "UserBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "User access token from POST /auth/login"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-based).",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "PageSize": {
        "name": "page_size",
        "in": "query",
        "description": "Items per page (max 100).",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
      "SearchQuery": {
        "name": "q",
        "in": "query",
        "description": "Search keywords.",
        "schema": { "type": "string" }
      },
      "Locale": {
        "name": "locale",
        "in": "query",
        "description": "Content locale when multilingual mode is enabled.",
        "schema": { "type": "string", "example": "en-US" }
      },
      "CategoryIDQuery": {
        "name": "category_id",
        "in": "query",
        "description": "Filter items assigned to this category ID.",
        "schema": { "type": "integer", "minimum": 1 }
      },
      "CategorySlugQuery": {
        "name": "category_slug",
        "in": "query",
        "description": "Filter items in this category slug and its descendants.",
        "schema": { "type": "string", "example": "news" }
      },
      "IncludeDescendants": {
        "name": "include_descendants",
        "in": "query",
        "description": "When used with category_id, include items in child categories.",
        "schema": { "type": "string", "enum": ["1"] }
      },
      "TagIDQuery": {
        "name": "tag_id",
        "in": "query",
        "description": "Filter items with this tag ID.",
        "schema": { "type": "integer", "minimum": 1 }
      },
      "Featured": {
        "name": "featured",
        "in": "query",
        "description": "Set to 1 to return featured items only.",
        "schema": { "type": "string", "enum": ["1"] }
      },
      "CategoriesFlat": {
        "name": "flat",
        "in": "query",
        "description": "Return a flat category list instead of a nested tree.",
        "schema": { "type": "string", "enum": ["1"] }
      },
      "CategoriesRoots": {
        "name": "roots",
        "in": "query",
        "description": "Return only top-level categories.",
        "schema": { "type": "string", "enum": ["1"] }
      },
      "CategoriesParentID": {
        "name": "parent_id",
        "in": "query",
        "description": "Return direct children of this parent category ID.",
        "schema": { "type": "integer", "minimum": 1 }
      },
...
    }
  },
  "security": [{ "ApiKeyAuth": [] }],
  "paths": {
...
    "/items": {
      "get": {
        "operationId": "listItems",
        "tags": ["Items"],
        "summary": "List visible published items",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" },
          { "$ref": "#/components/parameters/SearchQuery" },
          { "$ref": "#/components/parameters/Locale" },
          { "$ref": "#/components/parameters/CategoryIDQuery" },
          { "$ref": "#/components/parameters/CategorySlugQuery" },
          { "$ref": "#/components/parameters/IncludeDescendants" },
          { "$ref": "#/components/parameters/TagIDQuery" },
          { "$ref": "#/components/parameters/Featured" }
        ],
        "responses": { "200": { "description": "Paginated items" } }
      }
    },
...
    "/categories": {
      "get": {
        "operationId": "listCategories",
        "tags": ["Categories"],
        "summary": "Category tree",
        "description": "Returns a nested tree by default. Use query parameters to change shape.",
        "parameters": [
          { "$ref": "#/components/parameters/Locale" },
          { "$ref": "#/components/parameters/CategoriesFlat" },
          { "$ref": "#/components/parameters/CategoriesRoots" },
          { "$ref": "#/components/parameters/CategoriesParentID" }
        ],
        "responses": { "200": { "description": "Category tree or list" } }
      }
    },
...
  }
}